Address PR review comments (CORE-1293)
- Remove unused imports
- Remove unnecessary unboxing
- Add clarifying Javadoc on utility test code

+++ rdf-abac-core/src/main/java/io/telicent/jena/abac/labels/store/rocksdb/modern/DictionaryLabelStoreRocksDB.java
@@ -619,7 +619,7 @@ public class DictionaryLabelStoreRocksDB extends RocksDbLabelsStore implements L
                                     iterator.next();
                                     continue;
                                 }
-                                store.setLabel(newKey, newValue.longValue());
+                                store.setLabel(newKey, newValue);
                                 counter.incrementAndGet();
                                 batchCount++;
